requirements.txt
django
djangorestframework
pipwin
pillow
black
reportlab
requests
beautifulsoup4
pandas
pdbpp
prompt_toolkit
rich
matplotlib
folium
geopandas
descartes
shapely
contextily
mapclassify
djangorestframework-jsonapi
rest-pandas
django-cors-headers
django-pandas
messytables
